DAOS-16304 tools: Create libdaos_self_test
Move the non-UI logic from self_test.c into a new
library that can be used from other callers.

Features=CartSelfTest
Required-githooks: true
